pyproject.toml
[tool.poetry]
name = "automation-practice"
version = "0.1.0"
description = ""
authors = ["Austine Were <austinewuncler@gmail.com>"]
readme = "README.md"
packages = [{include = "automation_practice"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.9.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
